Final adjustments to gen_aig[e]fs_ics.py for production (#64)
This PR includes the last changes that SPA team made to
- `oper/gen_aigfs_ics.py`
- `oper/gen_aigefs_ics.py`

to operationalize the scripts. These changes are a band-aid fix that
will need a coordinated fix in the next release. A PR such as #54 would
likely resolve these issues. Some key areas we will need to improve on
include:
- `oper/gen_aigefs_ics.py`
- Remove boto initialization (caused failures during testing) and calls
to download data from s3 buckets
- Output netCDF file is still has `mlgefs` prefix, which is consistent
with
[exaigefs_prep.sh](https://github.com/NOAA-EMC/aigefs/blob/62f3cb50438e13fce5d95ddd3a09301a68ca1214/scripts/exaigefs_prep.sh#L40)
- Both scripts
  - Remove colon characters from intermediate file names
  - Explicitly locate files instead of looping and globbing (#36)